RpadUtil {Rpad}	R Documentation

Rpad utilities

Description

Rpad utilities to generate filenames or URL's

Usage

  RpadURL(filename = "")
  RpadBaseURL(filename = "")
  RpadBaseFile(filename = "")
  RpadIsLocal()

Arguments

filename 	 the name of a file.

Value

RpadURL returns the URL for the given filename: "./filename" for the local version of Rpad and "/Rpad/server/dd????????/filename" for the server version. Use this to output HTML links for the user.
RpadBaseURL returns the base URL: "filename" for the local version and "/Rpad/filename" for the client-server version. Use this to point the user to data files or other links on the server that is somewhere permanent. (The current R working directory is not permanent in the client-server version.)
RpadBaseFile returns the file name relative to the base R directory: "filename" for the local version and "../../filename" for the client-server version. Use this in R to read in data files or save data files somewhere permanent.

Author(s)

Tom Short, EPRI Solutions, Inc., (tshort@eprisolutions.com)

See Also

Rpad, RpadHTML

Examples

  # make some data
  x <- 1:10
  y2 <- x^3
  save(x, y2, file = RpadBaseFile("testdata.RData"))
  # output a link to the user:
  HTMLon()
  cat("<a href='", RpadBaseURL("testdata.RData"), sep="")
  cat("'>Click</a> to download the test data.")

[Package Rpad version 1.1.0 Index]